Deleting the rest of ex functions and files that have been already moved (the only file that left is dataset_test)
